% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{utf8_to_any}
\alias{utf8_to_any}
\title{Test Iconv}
\usage{
utf8_to_any(string, to = "latin1")
}
\arguments{
\item{string}{input string}

\item{string}{to iconv output encoding}
}
\description{
Test the RIconv C API.
}
